% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerialNumber.R
\name{SerialNumberServer}
\alias{SerialNumberServer}
\title{处理逻辑}
\usage{
SerialNumberServer(input, output, session, app_id, run_env = "PRD")
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}
}
\value{
返回值
}
\description{
处理逻辑
}
\examples{
SerialNumberServer()
}
